refactor: Extract settings discovery and parser from loader.rs
Split the 833-line loader.rs into three focused modules:
- discovery.rs: File discovery logic (config path resolution, loading from
  standard locations like user global, project, enterprise)
- parser.rs: Format parsing for JSON, TOML, and YAML config files
- loader.rs: Coordinator that orchestrates discovery + parsing into a
  SettingsHierarchy

Delegation methods on SettingsLoader are preserved for API stability.
All 49 settings tests pass with zero warnings.
